refactor: 일정 월별 조회 리팩토링
