+++ src/App.jsx
@@ -6,38 +6,37 @@ import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import './assets/fonts/fonts.css';
 
-import { ThemeProvider } from 'styled-components';
 import Attendance from '@/pages/Attendance';
+import AttendCheck from '@/pages/AttendCheck';
+import Board from '@/pages/Board';
 import Calendar from '@/pages/Calendar';
+import Dues from '@/pages/Dues';
+import Edit from '@/pages/Edit';
+import EventAdmin from '@/pages/EventAdmin';
+import EventDetails from '@/pages/EventDetails';
 import Home from '@/pages/Home';
 import Landing from '@/pages/Landing';
 import Login from '@/pages/Login';
-import Profile from '@/pages/Profile';
-import Signup from '@/pages/Signup';
-import EventDetails from '@/pages/EventDetails';
-import AttendCheck from '@/pages/AttendCheck';
 import Member from '@/pages/Member';
 import MemberDetail from '@/pages/MemberDetail';
 import MyPage from '@/pages/MyPage';
-import Edit from '@/pages/Edit';
-import Dues from '@/pages/Dues';
-import Board from '@/pages/Board';
-import StudyPosting from '@/pages/StudyPosting';
-import NoticePosting from '@/pages/NoticePosting';
-import EventAdmin from '@/pages/EventAdmin';
-import StudyDetail from '@/pages/StudyDetail';
 import NoticeDetail from '@/pages/NoticeDetail';
+import NoticePosting from '@/pages/NoticePosting';
+import Profile from '@/pages/Profile';
+import Signup from '@/pages/Signup';
+import StudyDetail from '@/pages/StudyDetail';
+import StudyPosting from '@/pages/StudyPosting';
+import { ThemeProvider } from 'styled-components';
 
-import theme from '@/styles/theme';
 import Receipt from '@/pages/Receipt';
+import theme from '@/styles/theme';
 
-import { UserProvider } from '@/api/UserContext';
-import { MonthlyScheduleProvider } from '@/api/MonthlyScheduleContext';
+import { AttendCheckProvider } from '@/api/AttendCheckContext';
 import { BoardProvider } from '@/api/BoardContext';
 import { DuesProvider } from '@/api/DuesContext';
 import { NoticeProvider } from '@/api/NoticeContext';
+import { UserProvider } from '@/api/UserContext';
 import { YearlyScheduleProvider } from '@/api/YearlyScheduleContext';
-import { AttendCheckProvider } from '@/api/AttendCheckContext';
 
 import ScrollToTop from '@/hooks/ScrollToTop';
 
@@ -45,46 +44,45 @@ const App = () => {
   return (
     <ThemeProvider theme={theme}>
       <UserProvider>
-        <MonthlyScheduleProvider>
-          <BoardProvider>
-            <DuesProvider>
-              <NoticeProvider>
-                <YearlyScheduleProvider>
-                  <AttendCheckProvider>
-                    <ScrollToTop />
-                    <Routes>
-                      <Route path="/" element={<Landing />} />
-                      <Route path="/login" element={<Login />} />
-                      <Route path="/attendance" element={<Attendance />} />
-                      <Route path="/calendar" element={<Calendar />} />
-                      <Route path="/:type/:id" element={<EventDetails />} />
-                      <Route path="/event/create" element={<EventAdmin />} />
-                      <Route path="/event/:id/edit" element={<EventAdmin />} />
-                      <Route path="/home" element={<Home />} />
-                      <Route path="/signup" element={<Signup />} />
-                      <Route path="/profile" element={<Profile />} />
-                      <Route path="/attendCheck" element={<AttendCheck />} />
-                      <Route path="/member" element={<Member />} />
-                      <Route path="/member/:id" element={<MemberDetail />} />
-                      <Route path="/mypage" element={<MyPage />} />
-                      <Route path="/edit" element={<Edit />} />
-                      <Route path="/dues" element={<Dues />} />
-                      <Route path="/receipt" element={<Receipt />} />
-                      <Route path="/board" element={<Board />} />
-                      <Route path="/study/:id" element={<StudyDetail />} />
-                      <Route path="/notice/:id" element={<NoticeDetail />} />
-                      <Route path="/study/post" element={<StudyPosting />} />
-                      <Route path="/notice/post" element={<NoticePosting />} />
-                    </Routes>
-                  </AttendCheckProvider>
-                </YearlyScheduleProvider>
-              </NoticeProvider>
-            </DuesProvider>
-          </BoardProvider>
-        </MonthlyScheduleProvider>
+        <BoardProvider>
+          <DuesProvider>
+            <NoticeProvider>
+              <YearlyScheduleProvider>
+                <AttendCheckProvider>
+                  <ScrollToTop />
+                  <Routes>
+                    <Route path="/" element={<Landing />} />
+                    <Route path="/login" element={<Login />} />
+                    <Route path="/attendance" element={<Attendance />} />
+                    <Route path="/calendar" element={<Calendar />} />
+                    <Route path="/:type/:id" element={<EventDetails />} />
+                    <Route path="/event/create" element={<EventAdmin />} />
+                    <Route path="/event/:id/edit" element={<EventAdmin />} />
+                    <Route path="/home" element={<Home />} />
+                    <Route path="/signup" element={<Signup />} />
+                    <Route path="/profile" element={<Profile />} />
+                    <Route path="/attendCheck" element={<AttendCheck />} />
+                    <Route path="/member" element={<Member />} />
+                    <Route path="/member/:id" element={<MemberDetail />} />
+                    <Route path="/mypage" element={<MyPage />} />
+                    <Route path="/edit" element={<Edit />} />
+                    <Route path="/dues" element={<Dues />} />
+                    <Route path="/receipt" element={<Receipt />} />
+                    <Route path="/board" element={<Board />} />
+                    <Route path="/study/:id" element={<StudyDetail />} />
+                    <Route path="/notice/:id" element={<NoticeDetail />} />
+                    <Route path="/study/post" element={<StudyPosting />} />
+                    <Route path="/notice/post" element={<NoticePosting />} />
+                  </Routes>
+                </AttendCheckProvider>
+              </YearlyScheduleProvider>
+            </NoticeProvider>
+          </DuesProvider>
+        </BoardProvider>
       </UserProvider>
     </ThemeProvider>
   );
 };
 
 export default App;
+
